move error types that are included in APIs into v1
