+++ oximeter/types/versions/src/impls/quantile.rs
@@ -7,28 +7,7 @@
 use crate::latest::histogram::HistogramSupport;
 use crate::latest::quantile::FILLED_MARKER_LEN;
 use crate::latest::quantile::Quantile;
-use schemars::JsonSchema;
-use serde::Deserialize;
-use serde::Serialize;
-use thiserror::Error;
-
-/// Errors related to constructing a `Quantile` instance or estimating the
-/// p-quantile.
-#[derive(
-    Debug, Clone, Error, PartialEq, JsonSchema, Serialize, Deserialize,
-)]
-#[serde(tag = "type", content = "content", rename_all = "snake_case")]
-pub enum QuantileError {
-    /// The p value must be in the range [0, 1].
-    #[error("The p value must be in the range [0, 1].")]
-    InvalidPValue,
-    /// Quantile estimation is not possible without samples.
-    #[error("Quantile estimation is not possible without any samples.")]
-    InsufficientSampleSize,
-    /// A non-finite was encountered, either as a bin edge or a sample.
-    #[error("Samples must be finite values, not Infinity or NaN.")]
-    NonFiniteValue,
-}
+use crate::latest::quantile::QuantileError;
 
 impl Quantile {
     /// Create a new `Quantile` instance.
