+++ oximeter/types/versions/src/initial/histogram.rs
@@ -5,12 +5,78 @@
 //! Types for managing metrics that are histograms.
 
 use super::quantile::Quantile;
+use super::quantile::QuantileError;
 use crate::impls::histogram::HistogramSupport;
 use chrono::DateTime;
 use chrono::Utc;
 use schemars::JsonSchema;
 use serde::Deserialize;
 use serde::Serialize;
+
+/// Errors related to constructing histograms or adding samples into them.
+#[derive(
+    Debug, Clone, thiserror::Error, JsonSchema, Serialize, Deserialize,
+)]
+#[serde(tag = "type", content = "content", rename_all = "snake_case")]
+pub enum HistogramError {
+    /// An attempt to construct a histogram with an empty set of bins.
+    #[error("Bins may not be empty")]
+    EmptyBins,
+
+    /// An attempt to construct a histogram with non-monotonic bins.
+    #[error("Bins must be monotonically increasing")]
+    NonmonotonicBins,
+
+    /// A non-finite was encountered, either as a bin edge or a sample.
+    #[error("Bin edges and samples must be finite values, not Infinity or NaN")]
+    NonFiniteValue,
+
+    /// Error returned when two neighboring bins are not adjoining (there's space between them)
+    #[error("Neigboring bins {left} and {right} are not adjoining")]
+    NonAdjoiningBins { left: String, right: String },
+
+    /// Bin and count arrays are of different sizes.
+    #[error(
+        "Bin and count arrays must have the same size, found {n_bins} and {n_counts}"
+    )]
+    ArraySizeMismatch { n_bins: usize, n_counts: usize },
+
+    /// Error returned when a quantization error occurs.
+    #[error("Quantization error")]
+    Quantization(#[from] QuantizationError),
+
+    /// Error returned when a quantile error occurs.
+    #[error("Quantile error")]
+    Quantile(#[from] QuantileError),
+}
+
+/// Errors occurring during quantizated bin generation.
+#[derive(
+    Clone, Debug, thiserror::Error, JsonSchema, Serialize, Deserialize,
+)]
+#[serde(tag = "type", content = "content", rename_all = "snake_case")]
+pub enum QuantizationError {
+    #[error("Overflow during bin generation")]
+    Overflow,
+
+    #[error("Precision error during bin generation")]
+    Precision,
+
+    #[error("Base must in the range [1, 32]")]
+    InvalidBase,
+
+    #[error("Number of steps must be > 1 and fit in the output type")]
+    InvalidSteps,
+
+    #[error(
+        "Number of steps must be multiple of base and \
+        evenly divide a power of the base"
+    )]
+    UnevenStepsForBase,
+
+    #[error("Low power must be strictly less than high power")]
+    PowersOutOfOrder,
+}
 
 /// A type storing a range over `T`.
 ///
